% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\docType{data}
\name{icons}
\alias{icons}
\title{Font Awesome icons tibble}
\format{tibble with columns
- name
- unicode
- horiz_adv_x
- d
- rune}
\usage{
icons
}
\description{
Font Awesome icons tibble
}
\keyword{datasets}
